graphite-client: fix error shown in error
The first flush always falls into the loggingMessageDisplayed condition
at least once before getting itself set properly.  Added a check to
require isLogging to also be false for this condition to be met.  This
prevents the message that logging is off being displayed on startup when
logging is actually on.

Fixes: #8